+++ release/c/wuffs-unsupported-snapshot.c
@@ -8280,6 +8280,7 @@ struct wuffs_lzw__decoder__struct {
 
 // ---------------- Status Codes
 
+extern const char wuffs_gif__error__bad_lzw_code[];
 extern const char wuffs_gif__error__bad_extension_label[];
 extern const char wuffs_gif__error__bad_frame_size[];
 extern const char wuffs_gif__error__bad_graphic_control[];
@@ -8478,8 +8479,8 @@ struct wuffs_gif__decoder__struct {
     uint64_t f_metadata_io_position;
     bool f_quirks[7];
     bool f_delayed_num_decoded_frames;
-    bool f_previous_lzw_decode_ended_abruptly;
     bool f_seen_header;
+    bool f_ignored_but_affects_benchmarks;
     bool f_has_global_palette;
     uint8_t f_interlace;
     bool f_seen_num_animation_loops_value;
@@ -8503,6 +8504,19 @@ struct wuffs_gif__decoder__struct {
     uint64_t f_compressed_ri;
     uint64_t f_compressed_wi;
     wuffs_base__pixel_swizzler f_swizzler;
+    uint32_t f_lzw_pending_literal_width_plus_one;
+    uint32_t f_lzw_literal_width;
+    uint32_t f_lzw_clear_code;
+    uint32_t f_lzw_end_code;
+    uint32_t f_lzw_save_code;
+    uint32_t f_lzw_prev_code;
+    uint32_t f_lzw_width;
+    uint32_t f_lzw_bits;
+    uint32_t f_lzw_n_bits;
+    uint32_t f_lzw_output_ri;
+    uint32_t f_lzw_output_wi;
+    uint32_t f_lzw_read_from_return_value;
+    uint16_t f_lzw_prefixes[4096];
 
     uint32_t p_decode_image_config[1];
     uint32_t p_do_decode_image_config[1];
@@ -8529,7 +8543,9 @@ struct wuffs_gif__decoder__struct {
     uint8_t f_compressed[4096];
     uint8_t f_palettes[2][1024];
     uint8_t f_dst_palette[1024];
-    wuffs_lzw__decoder f_lzw;
+    uint8_t f_lzw_suffixes[4096][8];
+    uint16_t f_lzw_lm1s[4096];
+    uint8_t f_lzw_output[8199];
 
     struct {
       uint32_t v_background_color;
@@ -35590,6 +35606,7 @@ wuffs_lzw__decoder__flush(
 
 // ---------------- Status Codes Implementations
 
+const char wuffs_gif__error__bad_lzw_code[] = "#gif: bad LZW code";
 const char wuffs_gif__error__bad_extension_label[] = "#gif: bad extension label";
 const char wuffs_gif__error__bad_frame_size[] = "#gif: bad frame size";
 const char wuffs_gif__error__bad_graphic_control[] = "#gif: bad graphic control";
@@ -35597,7 +35614,7 @@ const char wuffs_gif__error__bad_header[] = "#gif: bad header";
 const char wuffs_gif__error__bad_literal_width[] = "#gif: bad literal width";
 const char wuffs_gif__error__bad_palette[] = "#gif: bad palette";
 const char wuffs_gif__error__truncated_input[] = "#gif: truncated input";
-const char wuffs_gif__error__internal_error_inconsistent_ri_wi[] = "#gif: internal error: inconsistent ri/wi";
+const char wuffs_gif__error__internal_error_inconsistent_i_o[] = "#gif: internal error: inconsistent I/O";
 
 // ---------------- Private Consts
 
@@ -35762,6 +35779,17 @@ wuffs_gif__decoder__copy_to_image_buffer(
     wuffs_base__pixel_buffer* a_pb,
     wuffs_base__slice_u8 a_src);
 
+WUFFS_BASE__GENERATED_C_CODE
+static wuffs_base__empty_struct
+wuffs_gif__decoder__lzw_init(
+    wuffs_gif__decoder* self);
+
+WUFFS_BASE__GENERATED_C_CODE
+static wuffs_base__empty_struct
+wuffs_gif__decoder__lzw_read_from(
+    wuffs_gif__decoder* self,
+    wuffs_base__io_buffer* a_src);
+
 // ---------------- VTables
 
 const wuffs_base__image_decoder__func_ptrs
@@ -35841,13 +35869,6 @@ wuffs_gif__decoder__initialize(
     }
   }
 
-  {
-    wuffs_base__status z = wuffs_lzw__decoder__initialize(
-        &self->private_data.f_lzw, sizeof(self->private_data.f_lzw), WUFFS_VERSION, options);
-    if (z.repr) {
-      return z;
-    }
-  }
   self->private_impl.magic = WUFFS_BASE__MAGIC;
   self->private_impl.vtable_for__wuffs_base__image_decoder.vtable_name =
       wuffs_base__image_decoder__vtable_name;
@@ -38058,9 +38079,7 @@ wuffs_gif__decoder__decode_id_part1(
       }
       goto ok;
     }
-    if (self->private_impl.f_previous_lzw_decode_ended_abruptly) {
-      wuffs_base__ignore_status(wuffs_lzw__decoder__initialize(&self->private_data.f_lzw,
-          sizeof (wuffs_lzw__decoder), WUFFS_VERSION, WUFFS_INITIALIZE__LEAVE_INTERNAL_BUFFERS_UNINITIALIZED));
+    if (self->private_impl.f_ignored_but_affects_benchmarks) {
     }
     {
       WUFFS_BASE__COROUTINE_SUSPENSION_POINT(4);
@@ -38075,8 +38094,8 @@ wuffs_gif__decoder__decode_id_part1(
       status = wuffs_base__make_status(wuffs_gif__error__bad_literal_width);
       goto exit;
     }
-    wuffs_lzw__decoder__set_quirk(&self->private_data.f_lzw, 1348378624u, ((uint64_t)((1u + v_lw))));
-    self->private_impl.f_previous_lzw_decode_ended_abruptly = true;
+    self->private_impl.f_lzw_pending_literal_width_plus_one = ((uint32_t)((1u + v_lw)));
+    self->private_impl.f_ignored_but_affects_benchmarks = true;
 
     ok:
     self->private_impl.p_decode_id_part1[0] = 0;
@@ -38108,8 +38127,6 @@ wuffs_gif__decoder__decode_id_part2(
     wuffs_base__pixel_buffer* a_dst,
     wuffs_base__io_buffer* a_src,
     wuffs_base__slice_u8 a_workbuf) {
-  wuffs_base__io_buffer empty_io_buffer = wuffs_base__empty_io_buffer();
-
   wuffs_base__status status = wuffs_base__make_status(NULL);
 
   uint64_t v_block_size = 0;
@@ -38123,9 +38140,7 @@ wuffs_gif__decoder__decode_id_part2(
   const uint8_t* io1_v_r WUFFS_BASE__POTENTIALLY_UNUSED = NULL;
   const uint8_t* io2_v_r WUFFS_BASE__POTENTIALLY_UNUSED = NULL;
   uint64_t v_mark = 0;
-  wuffs_base__status v_lzw_status = wuffs_base__make_status(NULL);
   wuffs_base__status v_copy_status = wuffs_base__make_status(NULL);
-  wuffs_base__slice_u8 v_uncompressed = {0};
 
   const uint8_t* iop_a_src = NULL;
   const uint8_t* io0_a_src WUFFS_BASE__POTENTIALLY_UNUSED = NULL;
@@ -38146,6 +38161,7 @@ wuffs_gif__decoder__decode_id_part2(
   switch (coro_susp_point) {
     WUFFS_BASE__COROUTINE_SUSPENSION_POINT_0;
 
+    wuffs_gif__decoder__lzw_init(self);
     v_need_block_size = true;
     label__outer__continue:;
     while (true) {
@@ -38193,7 +38209,7 @@ wuffs_gif__decoder__decode_id_part2(
       }
       while (true) {
         if ((self->private_impl.f_compressed_ri > self->private_impl.f_compressed_wi) || (self->private_impl.f_compressed_wi > 4096u)) {
-          status = wuffs_base__make_status(wuffs_gif__error__internal_error_inconsistent_ri_wi);
+          status = wuffs_base__make_status(wuffs_gif__error__internal_error_inconsistent_i_o);
           goto exit;
         }
         {
@@ -38213,12 +38229,9 @@ wuffs_gif__decoder__decode_id_part2(
               self->private_impl.f_compressed_wi),
               0u);
           v_mark = ((uint64_t)(iop_v_r - io0_v_r));
-          {
-            u_r.meta.ri = ((size_t)(iop_v_r - u_r.data.ptr));
-            wuffs_base__status t_1 = wuffs_lzw__decoder__transform_io(&self->private_data.f_lzw, &empty_io_buffer, v_r, wuffs_base__utility__empty_slice_u8());
-            v_lzw_status = t_1;
-            iop_v_r = u_r.data.ptr + u_r.meta.ri;
-          }
+          u_r.meta.ri = ((size_t)(iop_v_r - u_r.data.ptr));
+          wuffs_gif__decoder__lzw_read_from(self, v_r);
+          iop_v_r = u_r.data.ptr + u_r.meta.ri;
           wuffs_base__u64__sat_add_indirect(&self->private_impl.f_compressed_ri, wuffs_base__io__count_since(v_mark, ((uint64_t)(iop_v_r - io0_v_r))));
           v_r = o_0_v_r;
           iop_v_r = o_0_iop_v_r;
@@ -38226,16 +38239,19 @@ wuffs_gif__decoder__decode_id_part2(
           io1_v_r = o_0_io1_v_r;
           io2_v_r = o_0_io2_v_r;
         }
-        v_uncompressed = wuffs_lzw__decoder__flush(&self->private_data.f_lzw);
-        if (((uint64_t)(v_uncompressed.len)) > 0u) {
-          v_copy_status = wuffs_gif__decoder__copy_to_image_buffer(self, a_dst, v_uncompressed);
+        if (self->private_impl.f_lzw_output_ri < self->private_impl.f_lzw_output_wi) {
+          v_copy_status = wuffs_gif__decoder__copy_to_image_buffer(self, a_dst, wuffs_base__make_slice_u8_ij(self->private_data.f_lzw_output,
+              self->private_impl.f_lzw_output_ri,
+              self->private_impl.f_lzw_output_wi));
           if (wuffs_base__status__is_error(&v_copy_status)) {
             status = v_copy_status;
             goto exit;
           }
+          self->private_impl.f_lzw_output_ri = 0u;
+          self->private_impl.f_lzw_output_wi = 0u;
         }
-        if (wuffs_base__status__is_ok(&v_lzw_status)) {
-          self->private_impl.f_previous_lzw_decode_ended_abruptly = false;
+        if (self->private_impl.f_lzw_read_from_return_value == 0u) {
+          self->private_impl.f_ignored_but_affects_benchmarks = false;
           if (v_need_block_size || (v_block_size > 0u)) {
             self->private_data.s_decode_id_part2[0].scratch = ((uint32_t)(v_block_size));
             WUFFS_BASE__COROUTINE_SUSPENSION_POINT(3);
@@ -38259,10 +38275,10 @@ wuffs_gif__decoder__decode_id_part2(
             }
           }
           goto label__outer__break;
-        } else if (v_lzw_status.repr == wuffs_base__suspension__short_read) {
-          goto label__outer__continue;
-        } else if (v_lzw_status.repr == wuffs_base__suspension__short_write) {
+        } else if (self->private_impl.f_lzw_read_from_return_value == 1u) {
           continue;
+        } else if (self->private_impl.f_lzw_read_from_return_value == 2u) {
+          goto label__outer__continue;
         } else if (self->private_impl.f_quirks[3u] && (self->private_impl.f_dst_y >= self->private_impl.f_frame_rect_y1) && (self->private_impl.f_interlace == 0u)) {
           if (v_need_block_size || (v_block_size > 0u)) {
             self->private_data.s_decode_id_part2[0].scratch = ((uint32_t)(v_block_size));
@@ -38287,15 +38303,15 @@ wuffs_gif__decoder__decode_id_part2(
             }
           }
           goto label__outer__break;
-        }
-        status = v_lzw_status;
-        if (wuffs_base__status__is_error(&status)) {
+        } else if (self->private_impl.f_lzw_read_from_return_value == 3u) {
+          status = wuffs_base__make_status(wuffs_gif__error__truncated_input);
           goto exit;
-        } else if (wuffs_base__status__is_suspension(&status)) {
-          status = wuffs_base__make_status(wuffs_base__error__cannot_return_a_suspension);
+        } else if (self->private_impl.f_lzw_read_from_return_value == 4u) {
+          status = wuffs_base__make_status(wuffs_gif__error__bad_lzw_code);
           goto exit;
         }
-        goto ok;
+        status = wuffs_base__make_status(wuffs_gif__error__internal_error_inconsistent_i_o);
+        goto exit;
       }
     }
     label__outer__break:;
@@ -38420,7 +38436,7 @@ wuffs_gif__decoder__copy_to_image_buffer(
     if (((uint64_t)(a_src.len)) == v_src_ri) {
       break;
     } else if (((uint64_t)(a_src.len)) < v_src_ri) {
-      return wuffs_base__make_status(wuffs_gif__error__internal_error_inconsistent_ri_wi);
+      return wuffs_base__make_status(wuffs_gif__error__internal_error_inconsistent_i_o);
     }
     v_n = ((uint64_t)((self->private_impl.f_frame_rect_x1 - self->private_impl.f_dst_x)));
     v_n = wuffs_base__u64__min(v_n, (((uint64_t)(a_src.len)) - v_src_ri));
@@ -38443,11 +38459,222 @@ wuffs_gif__decoder__copy_to_image_buffer(
       continue;
     }
     if (v_src_ri != ((uint64_t)(a_src.len))) {
-      return wuffs_base__make_status(wuffs_gif__error__internal_error_inconsistent_ri_wi);
+      return wuffs_base__make_status(wuffs_gif__error__internal_error_inconsistent_i_o);
     }
     break;
   }
   return wuffs_base__make_status(NULL);
+}
+
+// -------- func gif.decoder.lzw_init
+
+WUFFS_BASE__GENERATED_C_CODE
+static wuffs_base__empty_struct
+wuffs_gif__decoder__lzw_init(
+    wuffs_gif__decoder* self) {
+  uint32_t v_i = 0;
+
+  self->private_impl.f_lzw_literal_width = 8u;
+  if (self->private_impl.f_lzw_pending_literal_width_plus_one > 0u) {
+    self->private_impl.f_lzw_literal_width = (self->private_impl.f_lzw_pending_literal_width_plus_one - 1u);
+  }
+  self->private_impl.f_lzw_clear_code = (((uint32_t)(1u)) << self->private_impl.f_lzw_literal_width);
+  self->private_impl.f_lzw_end_code = (self->private_impl.f_lzw_clear_code + 1u);
+  self->private_impl.f_lzw_save_code = self->private_impl.f_lzw_end_code;
+  self->private_impl.f_lzw_prev_code = self->private_impl.f_lzw_end_code;
+  self->private_impl.f_lzw_width = (self->private_impl.f_lzw_literal_width + 1u);
+  self->private_impl.f_lzw_bits = 0u;
+  self->private_impl.f_lzw_n_bits = 0u;
+  self->private_impl.f_lzw_output_ri = 0u;
+  self->private_impl.f_lzw_output_wi = 0u;
+  v_i = 0u;
+  while (v_i < self->private_impl.f_lzw_clear_code) {
+    self->private_data.f_lzw_lm1s[v_i] = 0u;
+    self->private_data.f_lzw_suffixes[v_i][0u] = ((uint8_t)(v_i));
+    v_i += 1u;
+  }
+  return wuffs_base__make_empty_struct();
+}
+
+// -------- func gif.decoder.lzw_read_from
+
+WUFFS_BASE__GENERATED_C_CODE
+static wuffs_base__empty_struct
+wuffs_gif__decoder__lzw_read_from(
+    wuffs_gif__decoder* self,
+    wuffs_base__io_buffer* a_src) {
+  uint32_t v_clear_code = 0;
+  uint32_t v_end_code = 0;
+  uint32_t v_save_code = 0;
+  uint32_t v_prev_code = 0;
+  uint32_t v_width = 0;
+  uint32_t v_bits = 0;
+  uint32_t v_n_bits = 0;
+  uint32_t v_output_wi = 0;
+  uint32_t v_code = 0;
+  uint32_t v_c = 0;
+  uint32_t v_o = 0;
+  uint32_t v_steps = 0;
+  uint8_t v_first_byte = 0;
+  uint16_t v_lm1_b = 0;
+  uint16_t v_lm1_a = 0;
+
+  const uint8_t* iop_a_src = NULL;
+  const uint8_t* io0_a_src WUFFS_BASE__POTENTIALLY_UNUSED = NULL;
+  const uint8_t* io1_a_src WUFFS_BASE__POTENTIALLY_UNUSED = NULL;
+  const uint8_t* io2_a_src WUFFS_BASE__POTENTIALLY_UNUSED = NULL;
+  if (a_src && a_src->data.ptr) {
+    io0_a_src = a_src->data.ptr;
+    io1_a_src = io0_a_src + a_src->meta.ri;
+    iop_a_src = io1_a_src;
+    io2_a_src = io0_a_src + a_src->meta.wi;
+  }
+
+  v_clear_code = self->private_impl.f_lzw_clear_code;
+  v_end_code = self->private_impl.f_lzw_end_code;
+  v_save_code = self->private_impl.f_lzw_save_code;
+  v_prev_code = self->private_impl.f_lzw_prev_code;
+  v_width = self->private_impl.f_lzw_width;
+  v_bits = self->private_impl.f_lzw_bits;
+  v_n_bits = self->private_impl.f_lzw_n_bits;
+  v_output_wi = self->private_impl.f_lzw_output_wi;
+  while (true) {
+    if (v_n_bits < v_width) {
+      if (((uint64_t)(io2_a_src - iop_a_src)) >= 4u) {
+        v_bits |= ((uint32_t)(wuffs_base__peek_u32le__no_bounds_check(iop_a_src) << v_n_bits));
+        iop_a_src += ((31u - v_n_bits) >> 3u);
+        v_n_bits |= 24u;
+      } else if (((uint64_t)(io2_a_src - iop_a_src)) <= 0u) {
+        if (a_src && a_src->meta.closed) {
+          self->private_impl.f_lzw_read_from_return_value = 3u;
+        } else {
+          self->private_impl.f_lzw_read_from_return_value = 2u;
+        }
+        break;
+      } else {
+        v_bits |= (((uint32_t)(wuffs_base__peek_u8be__no_bounds_check(iop_a_src))) << v_n_bits);
+        iop_a_src += 1u;
+        v_n_bits += 8u;
+        if (v_n_bits >= v_width) {
+        } else if (((uint64_t)(io2_a_src - iop_a_src)) <= 0u) {
+          if (a_src && a_src->meta.closed) {
+            self->private_impl.f_lzw_read_from_return_value = 3u;
+          } else {
+            self->private_impl.f_lzw_read_from_return_value = 2u;
+          }
+          break;
+        } else {
+          v_bits |= (((uint32_t)(wuffs_base__peek_u8be__no_bounds_check(iop_a_src))) << v_n_bits);
+          iop_a_src += 1u;
+          v_n_bits += 8u;
+          if (v_n_bits < v_width) {
+            self->private_impl.f_lzw_read_from_return_value = 5u;
+            break;
+          }
+        }
+      }
+    }
+    v_code = ((v_bits) & WUFFS_BASE__LOW_BITS_MASK__U32(v_width));
+    v_bits >>= v_width;
+    v_n_bits -= v_width;
+    if (v_code < v_clear_code) {
+      self->private_data.f_lzw_output[v_output_wi] = ((uint8_t)(v_code));
+      v_output_wi = ((v_output_wi + 1u) & 8191u);
+      if (v_save_code <= 4095u) {
+        v_lm1_a = (((uint16_t)(self->private_data.f_lzw_lm1s[v_prev_code] + 1u)) & 4095u);
+        self->private_data.f_lzw_lm1s[v_save_code] = v_lm1_a;
+        if ((v_lm1_a % 8u) != 0u) {
+          self->private_impl.f_lzw_prefixes[v_save_code] = self->private_impl.f_lzw_prefixes[v_prev_code];
+          memcpy(self->private_data.f_lzw_suffixes[v_save_code],self->private_data.f_lzw_suffixes[v_prev_code], sizeof(self->private_data.f_lzw_suffixes[v_save_code]));
+          self->private_data.f_lzw_suffixes[v_save_code][(v_lm1_a % 8u)] = ((uint8_t)(v_code));
+        } else {
+          self->private_impl.f_lzw_prefixes[v_save_code] = ((uint16_t)(v_prev_code));
+          self->private_data.f_lzw_suffixes[v_save_code][0u] = ((uint8_t)(v_code));
+        }
+        v_save_code += 1u;
+        if (v_width < 12u) {
+          v_width += (1u & (v_save_code >> v_width));
+        }
+        v_prev_code = v_code;
+      }
+    } else if (v_code <= v_end_code) {
+      if (v_code == v_end_code) {
+        self->private_impl.f_lzw_read_from_return_value = 0u;
+        break;
+      }
+      v_save_code = v_end_code;
+      v_prev_code = v_end_code;
+      v_width = (self->private_impl.f_lzw_literal_width + 1u);
+    } else if (v_code <= v_save_code) {
+      v_c = v_code;
+      if (v_code == v_save_code) {
+        v_c = v_prev_code;
+      }
+      v_o = ((v_output_wi + (((uint32_t)(self->private_data.f_lzw_lm1s[v_c])) & 4294967288u)) & 8191u);
+      v_output_wi = ((v_output_wi + 1u + ((uint32_t)(self->private_data.f_lzw_lm1s[v_c]))) & 8191u);
+      v_steps = (((uint32_t)(self->private_data.f_lzw_lm1s[v_c])) >> 3u);
+      while (true) {
+        memcpy((self->private_data.f_lzw_output)+(v_o), (self->private_data.f_lzw_suffixes[v_c]), 8u);
+        if (v_steps <= 0u) {
+          break;
+        }
+        v_steps -= 1u;
+        v_o = (((uint32_t)(v_o - 8u)) & 8191u);
+        v_c = ((uint32_t)(self->private_impl.f_lzw_prefixes[v_c]));
+      }
+      v_first_byte = self->private_data.f_lzw_suffixes[v_c][0u];
+      if (v_code == v_save_code) {
+        self->private_data.f_lzw_output[v_output_wi] = v_first_byte;
+        v_output_wi = ((v_output_wi + 1u) & 8191u);
+      }
+      if (v_save_code <= 4095u) {
+        v_lm1_b = (((uint16_t)(self->private_data.f_lzw_lm1s[v_prev_code] + 1u)) & 4095u);
+        self->private_data.f_lzw_lm1s[v_save_code] = v_lm1_b;
+        if ((v_lm1_b % 8u) != 0u) {
+          self->private_impl.f_lzw_prefixes[v_save_code] = self->private_impl.f_lzw_prefixes[v_prev_code];
+          memcpy(self->private_data.f_lzw_suffixes[v_save_code],self->private_data.f_lzw_suffixes[v_prev_code], sizeof(self->private_data.f_lzw_suffixes[v_save_code]));
+          self->private_data.f_lzw_suffixes[v_save_code][(v_lm1_b % 8u)] = v_first_byte;
+        } else {
+          self->private_impl.f_lzw_prefixes[v_save_code] = ((uint16_t)(v_prev_code));
+          self->private_data.f_lzw_suffixes[v_save_code][0u] = ((uint8_t)(v_first_byte));
+        }
+        v_save_code += 1u;
+        if (v_width < 12u) {
+          v_width += (1u & (v_save_code >> v_width));
+        }
+        v_prev_code = v_code;
+      }
+    } else {
+      self->private_impl.f_lzw_read_from_return_value = 4u;
+      break;
+    }
+    if (v_output_wi > 4095u) {
+      self->private_impl.f_lzw_read_from_return_value = 1u;
+      break;
+    }
+  }
+  if (self->private_impl.f_lzw_read_from_return_value != 2u) {
+    while (v_n_bits >= 8u) {
+      v_n_bits -= 8u;
+      if (iop_a_src > io1_a_src) {
+        iop_a_src--;
+      } else {
+        self->private_impl.f_lzw_read_from_return_value = 5u;
+        break;
+      }
+    }
+  }
+  self->private_impl.f_lzw_save_code = v_save_code;
+  self->private_impl.f_lzw_prev_code = v_prev_code;
+  self->private_impl.f_lzw_width = v_width;
+  self->private_impl.f_lzw_bits = v_bits;
+  self->private_impl.f_lzw_n_bits = v_n_bits;
+  self->private_impl.f_lzw_output_wi = v_output_wi;
+  if (a_src && a_src->data.ptr) {
+    a_src->meta.ri = ((size_t)(iop_a_src - a_src->data.ptr));
+  }
+
+  return wuffs_base__make_empty_struct();
 }
 
 #endif  // !defined(WUFFS_CONFIG__MODULES) || defined(WUFFS_CONFIG__MODULE__GIF)