+++ script/print-image-metadata.cc
@@ -54,7 +54,6 @@ print-image-metadata prints images' metadata.
 #define WUFFS_CONFIG__MODULE__DEFLATE
 #define WUFFS_CONFIG__MODULE__GIF
 #define WUFFS_CONFIG__MODULE__JPEG
-#define WUFFS_CONFIG__MODULE__LZW
 #define WUFFS_CONFIG__MODULE__NETPBM
 #define WUFFS_CONFIG__MODULE__NIE
 #define WUFFS_CONFIG__MODULE__PNG